refactor: organize user and profile types
