+++ Frontend-nextjs/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./types/sidebar";
-export * from "./types/darkmode";
 export * from "./types/transaction";
 export * from "./types/category";
 export * from "./types/ricorrenza";
+export * from "./types/user";
+export * from "./profilo/row";
